% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congress_2016.R
\name{congress_2016}
\alias{congress_2016}
\title{제20대 2016년 국회의원 선거}
\format{
list-column을 갖는 티블
}
\source{
http://www.nec.go.kr/portal/bbs/view/B0000338/37352.do?menuNo=200185&searchYear=&searchMonth=&searchWrd=\%EA\%B0\%9C\%ED\%91\%9C\%EA\%B2\%B0\%EA\%B3\%BC&searchCnd=1&viewType=&pageIndex=1&section=&searchOption1=
}
\description{
제20대 2016년 국회의원 선거
}
\examples{
  <ec>냽珥덉떆 <- congress_2016 \%>\% filter(precinct == "<ec>냽珥덉떆") \%>\% pull(data_clean) \%>\% .[[1]]
}
