% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congress.R
\docType{data}
\name{congress_2018}
\alias{congress_2018}
\title{Local Election 2018 in Korea - Congressman}
\format{tibble with columns}
\source{
http://www.nec.go.kr/portal/bbs/view/B0000338/37352.do?menuNo=200185&searchYear=&searchMonth=&searchWrd=%EA%B0%9C%ED%91%9C%EA%B2%B0%EA%B3%BC&searchCnd=1&viewType=&pageIndex=1&section=&searchOption1=
}
\usage{
congress_2018
}
\description{
Local Election 2018 in Korea - Congressman
}
\examples{
  congress_2018
}
\keyword{datasets}
