% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congress_2020.R
\name{congress_2020}
\alias{congress_2020}
\title{제21대 2020년 국회의원 선거}
\format{
list-column을 갖는 티블
}
\source{
\url{https://www.nec.go.kr/site/nec/ex/bbs/View.do?cbIdx=1129&bcIdx=15052}
}
\description{
2020년 4월 15일 치뤄진 국회의원 선거 득표 데이터
}
\examples{
  <ec>냽珥덉떆 <- congress_2020 \%>\% filter(precinct == "<ec>냽珥덉떆") \%>\% pull(data_clean) \%>\% .[[1]]
  bundang_eul <- congress_2020 \%>\% filter(precinct == "<ec>꽦<eb>궓<ec>떆遺꾨떦援ъ쓣") \%>\% pull(data_clean) \%>\% .[[1]]
}
