% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_2021_by_election.R
\name{demographics_tbl}
\alias{demographics_tbl}
\title{행정안전부 연령별 인구현황}
\format{
리스트 칼럼을 갖는 티블
}
\source{
\url{https://jumin.mois.go.kr/}
}
\description{
2020년 12월 31일 기준 (2021년 1월 1일) 시도별, 연령별, 성별 인구수
}
\examples{
  demographics_tbl
}
