% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.R
\docType{data}
\name{local_2018}
\alias{local_2018}
\title{Local Election 2018 in Korea - Provincial Governor}
\format{
tibble with columns
}
\source{
http://www.nec.go.kr/portal/bbs/view/B0000338/38843.do?menuNo=200185&searchYear=&searchMonth=&searchWrd=\%EC\%A7\%80\%EB\%B0\%A9\%EC\%84\%A0\%EA\%B1\%B0&searchCnd=1&viewType=&pageIndex=1&section=&searchOption1=
}
\usage{
local_2018
}
\description{
Local Election 2018 in Korea - Provincial Governor
}
\examples{
  local_df <- local_2018
}
\keyword{datasets}
