% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_2018_sigungu.R
\name{local_2018_sigungu}
\alias{local_2018_sigungu}
\title{제7회 2018년 지방선거 - 구시군의 장}
\format{
tibble with columns
}
\source{
\url{https://www.nec.go.kr/site/nec/ex/bbs/View.do?cbIdx=1129&bcIdx=14979}
}
\description{
2018년 6월 13일 대한민국 전역에서 치러진 지방의회 의원 및 지방자치단체의 장을 선출하기 위한 선거
}
\examples{
  local_2018_sigungu
}
