% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_2021_by_election.R
\name{local_2021_by_election}
\alias{local_2021_by_election}
\title{2021년 지방 재보궐선거: 서울특별시/부산광역시}
\format{
리스트 칼럼을 갖는 티블
}
\source{
\url{http://info.nec.go.kr/}
}
\description{
2021년 4월 7일 총 21개 선거구에서 치뤄진 재보궐 선거다.
\itemize{
\item 광역단체장: 2명
\item 기초단체장: 2명
\item 광역의원: 8명
\item 기초의원: 9명
}
}
\examples{
  local_2021_by_election
}
