% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_2018_sigungu.R
\docType{data}
\name{local_sigungu_2018}
\alias{local_sigungu_2018}
\title{지방선거 2018년 - 시군구장}
\format{
tibble with columns
}
\source{
http://www.nec.go.kr/portal/bbs/view/B0000338/38843.do?menuNo=200185&searchYear=&searchMonth=&searchWrd=\%EC\%A7\%80\%EB\%B0\%A9\%EC\%84\%A0\%EA\%B1\%B0&searchCnd=1&viewType=&pageIndex=1&section=&searchOption1=
}
\usage{
local_sigungu_2018
}
\description{
지방선거 2018년 - 시군구장
}
\examples{
  local_2018
}
\keyword{datasets}
