% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precinct.R
\docType{data}
\name{precinct}
\alias{precinct}
\title{Korean Election Precinct 2018}
\format{
tibble with columns
}
\source{
http://www.nec.go.kr/portal/bbs/view/B0000338/32767.do?menuNo=200185
}
\usage{
precinct
}
\description{
Korean Election Precinct 2018
}
\examples{
  precinct_df <- precinct
}
\keyword{datasets}
