% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precinct_2016.R
\name{precinct_2016}
\alias{precinct_2016}
\title{국회의원 2016년 선거구}
\format{
tibble with columns
}
\source{
http://www.nec.go.kr/portal/bbs/view/B0000338/32767.do?menuNo=200185
}
\description{
국회의원 2016년 선거구
}
\examples{
  precinct_tbl <- precinct_2016
}
