% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/president.R
\docType{data}
\name{president}
\alias{president}
\title{Korean Presidential Election 2018}
\format{tibble with columns}
\source{
http://www.nec.go.kr/portal/bbs/view/B0000338/36213.do?menuNo=200061&searchYear=&searchMonth=&searchWrd=%EB%8C%80%ED%86%B5%EB%A0%B9&searchCnd=1&viewType=&pageIndex=1&section=&searchOption1=
}
\usage{
president
}
\description{
Korean Presidential Election 2018
}
\examples{
  president
}
\keyword{datasets}
