% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/president.R
\name{president_2017}
\alias{president_2017}
\title{제19대 대통령선거 2017년}
\format{
18,455 관측점과 22개 변수를 갖는 티블
}
\source{
http://www.nec.go.kr/portal/bbs/view/B0000338/36213.do?menuNo=200061&searchYear=&searchMonth=&searchWrd=\%EB\%8C\%80\%ED\%86\%B5\%EB\%A0\%B9&searchCnd=1&viewType=&pageIndex=1&section=&searchOption1=
}
\description{
제19대 대통령선거 2017년
}
\examples{
  president_tbl <- president_2017
}
