% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congress_2016.R
\name{congress_2016}
\alias{congress_2016}
\title{제20대 2016년 국회의원 선거}
\format{
list-column을 갖는 티블
}
\source{
\url{https://www.nec.go.kr/site/nec/ex/bbs/View.do?cbIdx=1129&bcIdx=14879}
}
\description{
2016년 4월 13일 치뤄진 국회의원 선거
}
\examples{
  <ec>냽珥덉떆 <- congress_2016 \%>\% filter(precinct == "<ec>냽珥덉떆") \%>\% pull(data_clean) \%>\% .[[1]]
}
